Move types into types module
