+++ frontend/src/metabase/core/components/AccordionList/AccordionList.tsx
@@ -20,59 +20,8 @@ import { Icon, type IconName } from "metabase/ui";
 
 import { AccordionListRoot } from "./AccordionList.styled";
 import { AccordionListCell } from "./AccordionListCell";
+import type { Item, Row, Section } from "./types";
 import { type Cursor, getNextCursor, getPrevCursor } from "./utils";
-
-export type Section<T extends Item = Item> = {
-  key?: string;
-  name?: ReactNode;
-  displayName?: ReactNode;
-  type?:
-    | "action"
-    | "header"
-    | "search"
-    | "loading"
-    | "no-results"
-    | "item"
-    | "back";
-  icon?: IconName | null;
-  loading?: boolean;
-  items?: T[];
-  active?: boolean;
-  className?: string | null;
-};
-
-export type Item = object;
-
-export type Row<T extends object> = {
-  section: Section<T>;
-  sectionIndex: number;
-  isLastSection: boolean;
-} & (
-  | {
-      type: "action";
-    }
-  | {
-      type: "header";
-    }
-  | {
-      type: "search";
-    }
-  | {
-      type: "loading";
-    }
-  | {
-      type: "no-results";
-    }
-  | {
-      type: "back";
-    }
-  | {
-      type: "item";
-      item: T;
-      itemIndex: number;
-      isLastItem: boolean;
-    }
-);
 
 type Props<T extends Item> = {
   style?: CSSProperties & {
